WIP Implementation of CSS Roster Changer
Should make it possible to switch between different versions of the roster through the code menu in P+ based builds that don't use the CSSRoster.dat system.

There is one bug though in that currently it's breaking Melee Random; the cursor just kinda flies off the screen and crashes the game most of the time.